Add putting processes to sleep
Current sys_wait was waiting in kernel for the end of given process.
It is quite obvious why it was a poor idea :)

New sys_wait puts a process to sleep and adds it to the wait_queue of
the process it's waiting for. On process exit processes in its
wait_queue are woken.

Minor fixes:

	- arch_kprocess_regs_init sets GS to the KERNEL_DS, so
	  functions copying from/to user will work if a kernel process
	  uses syscall,
	- seriald uses waitpid instead of sys_waitpid