Rework buffer and add wait_queue
Circular buffer implementation was inflexible and needed in-code
initialization by buffer_init which was allocating memory for the
table.

Now, buffer doesn't need any in-code initialization and can be created
by the DECLARE_BUFFER macro, which can create various length buffers
(up to 64k chars). wait_queue has been built on the top of the list
implementation and is quite flexible, but as long as there are no
spinlocks or semaphores it can be buggy.

Minor changes:

	- keyboard driver uses new buffer and wait_queue
	- serial driver also uses new buffer (wait_queue to be
	  implemented)