Fix kernel symbols reading
Kernel symbols table was just an array of pointers to the allocated
structures. Its size was limited.

I decided to first check the number of symbols and then to allocate
table for them. It is slower, but we are not limited to some constant
size of the array and we are not wasting space (when we found fewer
symbols than fixed array size).

Minor changes:

	- interrupts are enabled by the kmain, not the arch_setup