Fix paging
There were many problems with fully mapping kernel (and user, but
that's temporary) space into 0xc000000. The page directory was not
reloaded and it was possible to reach low addresses that had been
used earlier.

The page directory is reloaded before a call to the kmain. Also, a
limit of GDT entry for user data has been fixed (it was 0xffff, which
resulted in a stack segment fault). All directly given addresses have
been changed into virtual addresses (if necessary).

Minor changes:

	- fixed RESTORE_ALL macro
	- added syscall_regs_check, which is called on every syscall
	  and checks if the values of the segments registers are
	  correct